Make sure the original line location is shown
Currently always when a binding error occurs, then in the console the error occurs at line 1936 (or so) in knockout-2.2.1.debug.js
That sucks.
The original line location is masked out and being lost.

With this change we're displaying the same kockout-wish error message but preserve the original line location!

This is very handful, please merge the pull request.